[quantization] Load models with float32 dtype in examples (#523)
This commit loads models with float32 dtype in examples.

TICO-DCO-1.0-Signed-off-by: seongwoo <mhs4670go@naver.com>

+++ tico/quantization/wrapq/examples/llama/quantize_llama_attn.py
@@ -26,7 +26,7 @@ from tico.quantization.wrapq.wrappers.llama.quant_attn import QuantLlamaAttentio
 from tico.utils.utils import SuppressWarning
 
 name = "Maykeye/TinyLLama-v0"
-model = AutoModelForCausalLM.from_pretrained(name)
+model = AutoModelForCausalLM.from_pretrained(name, dtype=torch.float32)
 tokenizer = AutoTokenizer.from_pretrained(name)
 
 # -------------------------------------------------------------------------
